Merging backup with load-osm script.
For unknown reasons, the data import seems to pass but running the backup script shows that the role names are missing (which should have been created through the init script).
Merging these together so that a backup is made immediately after transforming the OSM data into Nominatim format in an attempt to avoid troubleshooting why splitting up the role creation from the data ingestion has had an effect of losing users and roles in the database.