Fixed export filename in load-osm.sh
When running load-osm.sh now, a new volume must be mounted in the Docker container when this script is run (externally through the -v flag to docker run).
The script now produces a backup of the data immediately after creation in a file named `pgdumpall.dump` located at /var/exports/ in the container.
Consult the changes to the readme for instructions on running it successfully.