Added chown to backup file in load-osm.sh
`load-osm.sh` now accepts two new arguments: numeric user id, and numeric group id, of a user on the system outside of the container to change ownership of the backup file to after creation.
Before this change, the owner would be the docker container's postgres user and group, which causes the host system to not be able to move or alter the resulting backup file.
After this change, the backup file will be owned by the uid/gid passed in, allowing a process controlling the build/deployment to move file (push to s3, delete when finished, etc).
Updated the readme with example call of obtaining the exact uid/gid in the load/backup section. Also removed the last step which did not belong and reworded it into the empty restore section.