Fixed incorrect docker volume mount instructions.
The previous instructions caused restoring a backup to happen on a postgres directory that was not live (after initialization postgres serves up data from /etc/lib/postgres instead of /var/lib/postgres ...), which was causing the restoration to occur in a container instead of on the mounted volume. This meant each backup was written to a layer in a docker container and never to the mount point on the host os.
This change fixes those instructions and also changes the restore.sh script to write the restoration to the live folder (argument taken away).