Delete PGDIR at start of init.sh.
`init.sh` will now destroy the /data/$PGDIR directory if it exists prior to attempting to initialize the database.
This was added to make the workflow of creating, backing up, and restoring a Nominatim database idempotent.
The init.sh step will now be used by Jenkins during a deployment to wipe away the old database, and set up for importing a new database from backup.
It will also be used during an off-prod-host build to create the data directory in preparation for loading up an OSM file.